Remove BaseObject interface.
