+++ vm/object_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 )
 
-func getBuiltInMethod(t *testing.T, receiver BaseObject, methodName string) builtinMethodBody {
+func getBuiltInMethod(t *testing.T, receiver Object, methodName string) builtinMethodBody {
 	m := receiver.returnClass().lookupInstanceMethod(methodName)
 
 	if m == nil {
